Split engine.py into a package; house the executor in its own module
engine.py had grown to ~1950 lines once the executor landed. Split it
into a kestrel/engine/ package so the executor — a kernel component —
lives in its own module rather than beside InferenceEngine:

- _types.py: value types (EngineResult/Metrics/Stream, request/result
  containers, Completion, TickResult) shared without an import cycle
- executor.py: the Executor protocol, admission coordinator, and
  AutoregressiveExecutor
- core.py: InferenceEngine (the kernel)
- __init__.py: re-exports, so every existing 'from kestrel.engine import
  ...' keeps working unchanged

Pure restructure: no behavior change. Full engine + scheduler suite
passes (121) and GPU end-to-end (single + concurrent decode) is
unchanged.
